# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/rules": {
      "get": {
        "description": "Returns all available axe-core accessibility rules with metadata, sorted alphabetically by ID.",
        "operationId": "listRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of axe-core rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all axe-core rules",
        "tags": [
          "Rules"
        ]
      }
    },
    "/v1/scans/{id}/reports/html": {
      "get": {
        "description": "Returns a full standalone HTML accessibility report.",
        "operationId": "getHtmlReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scan ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML accessibility report",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate an HTML accessibility report for a scan",
        "tags": [
          "Reports"
        ]
      }
    },
    "/v1/scans/{id}/reports/pdf": {
      "get": {
        "description": "Returns a generated PDF accessibility report.",
        "operationId": "getPdfReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scan ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF accessibility report",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate a PDF accessibility report for a scan",
        "tags": [
          "Reports"
        ]
      }
    },
    "/v1/cleanup": {
      "post": {
        "description": "Immediately runs the cleanup process to delete scans older than the configured retention period, bypassing the enabled flag.",
        "operationId": "triggerCleanup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cleanup completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Trigger manual cleanup",
        "tags": [
          "Cleanup"
        ]
      }
    },
    "/v1/cleanup/policy": {
      "get": {
        "description": "Returns the current cleanup retention policy derived from environment variables.",
        "operationId": "getCleanupPolicy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cleanup retention policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupConfigDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cleanup retention policy",
        "tags": [
          "Cleanup"
        ]
      }
    }
  },
  "info": {
    "title": "MindfulAPI",
    "description": "Automated web accessibility scanning API powered by axe-core and Playwright. Built for use with the [mindfula11y](https://github.com/crinis/mindfula11y) TYPO3 extension. Errors are returned as RFC 9457 `application/problem+json`.",
    "version": "0.5.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Set the AUTH_TOKEN environment variable to enable authentication. When authentication is disabled (AUTH_DISABLED=true), the token is ignored.",
        "type": "http"
      }
    },
    "schemas": {
...
      "RuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "color-contrast",
            "description": "Axe rule identifier."
          },
          "description": {
            "type": "string",
            "example": "Elements must have sufficient color contrast"
          },
          "helpUrl": {
            "type": "string",
            "example": "https://dequeuniversity.com/rules/axe/4.11/color-contrast?application=axeAPI",
            "description": "Link to remediation guidance on Deque University.",
            "format": "uri",
            "nullable": true
          },
          "tags": {
            "example": [
              "wcag2aa",
              "wcag143",
              "cat.color"
            ],
            "description": "WCAG and category tags for this rule.",
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "description",
          "helpUrl"
        ]
      },
...
      "ValidationErrorDetail": {
        "type": "object",
        "properties": {
          "pointer": {
            "type": "string",
            "example": "/urls",
            "description": "JSON Pointer to the field that failed validation."
          },
          "message": {
            "type": "string",
            "example": "urls must contain at least 2 elements",
            "description": "Human-readable validation message."
          }
        },
        "required": [
          "pointer",
          "message"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank",
            "description": "A URI reference identifying the problem type. `about:blank` is used for problems with no dedicated documentation."
          },
          "title": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "example": 400,
            "minimum": 400,
            "maximum": 599,
            "description": "HTTP status code for this occurrence."
          },
          "detail": {
            "type": "string",
            "example": "Scan with ID 42 not found",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "example": "/v1/scans/42",
            "description": "URI reference identifying this specific occurrence."
          },
          "errors": {
            "description": "Field-level validation errors (validation problems only).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
      "CleanupResultDto": {
        "type": "object",
        "properties": {
          "deletedScans": {
            "type": "integer",
            "example": 12,
            "description": "Number of scan runs deleted.",
            "minimum": 0
          },
          "cutoffDate": {
            "type": "string",
            "example": "2026-06-09T00:00:00.000Z",
            "description": "Scans created before this timestamp were deleted.",
            "format": "date-time"
          }
        },
        "required": [
          "deletedScans",
          "cutoffDate"
        ]
      },
      "CleanupConfigDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true,
            "description": "Whether scheduled automatic cleanup is enabled"
          },
          "retentionDays": {
            "type": "integer",
            "example": 30,
            "description": "Number of days to retain scans before deletion",
            "minimum": 1
          },
          "interval": {
            "type": "string",
            "example": "0 2 * * *",
            "description": "Cron expression controlling the cleanup schedule"
          }
        },
        "required": [
          "enabled",
          "retentionDays",
          "interval"
        ]
      }
    }
  },
  "security": [
    {
      "bearer": []
    },
    {}
  ]
}