# src/controllers/rules.controller.ts
import { Controller, Get } from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from '@nestjs/swagger';
import { RulesService } from '../services/rules.service';
import { RuleResponseDto } from '../dto/scan/response';
import { ApiProblemResponses } from '../decorators/api-problem-responses.decorator';

/**
 * REST API controller for axe-core accessibility rules.
 * @route /rules
 */
@ApiTags('Rules')
@ApiBearerAuth()
@Controller('rules')
export class RulesController {
  /**
   * @param rulesService Service that exposes axe rule metadata.
   */
  constructor(private readonly rulesService: RulesService) {}

  @Get()
  @ApiOperation({
    operationId: 'listRules',
    summary: 'List all axe-core rules',
    description:
      'Returns all available axe-core accessibility rules with metadata, sorted alphabetically by ID.',
  })
  @ApiResponse({
    status: 200,
    description: 'Array of axe-core rules',
    type: RuleResponseDto,
    isArray: true,
  })
  @ApiProblemResponses(401, 429, 500)
  /**
   * Lists all available axe-core rules with response metadata.
   */
  getRules(): RuleResponseDto[] {
    return this.rulesService.getRules();
  }
}